move to an extra package
