+++ ui/org.eclipse.pde.core/src/org/eclipse/pde/internal/core/sourcelocators/EclipsePluginSourcePathLocator.java
@@ -11,7 +11,7 @@
  * Contributors:
  *     Christoph Läubrich - initial API and implementation
  *******************************************************************************/
-package org.eclipse.pde.internal.core;
+package org.eclipse.pde.internal.core.sourcelocators;
 
 import java.io.File;
 import org.eclipse.core.runtime.IPath;
